now with an infinite loop in main

+++ src/main/scala/traficSimulator/CarTheater/CreationActor.scala
@@ -19,7 +19,7 @@ class CreationActor(private val carsHolder : ActorRef) extends Actor{
   val log = Logging(context.system, this)
 
   def receive = {
-    case creationTimestamp : BigInt => {
+    case creationTimestamp : Double  => {
       val carId = generateUniqueId
       val carProps : Props = Props(classOf[CarActor],
           RouteApiWrapper.returnRoute(
